+++ src/tools/layer2/index.ts
@@ -1,6 +1,7 @@
 import type { ToolRegistrar } from "../types.js";
 import { registerAnimateParameter } from "./animateParameter.js";
 import { registerArrangeNetwork } from "./arrangeNetwork.js";
+import { registerBindToChannel } from "./bindToChannel.js";
 import { registerConnectNodes } from "./connectNodes.js";
 import { registerCreateContainer } from "./createContainer.js";
 import { registerCreateControlPanel } from "./createControlPanel.js";
@@ -23,6 +24,7 @@ export const layer2Registrars: ToolRegistrar[] = [
   registerCreateContainer,
   registerCreateControlPanel,
   registerAnimateParameter,
+  registerBindToChannel,
   registerManagePresets,
   registerManageCheckpoint,
   registerManageComponent,